Define 'header' and 'information' sections
First of all, we listed the different subsections of the main areas of
the HTML file. Then, we started working on the style. We mainly focus on
logo, menu, hero, and information sub-sections.
Pay attention: you forgot to separate the work in littler parts, missing
occasion to commit.
Stay more focused and be more consistent in committing more often.